Refactor Header: Кнопка добавления фильма вынесена в компонент
