+++ app/components/AddButton/AddButton.module.scss
@@ -1,0 +1,69 @@
+@import "styles/mixins.scss";
+
+.root {
+  position: relative;
+  min-width: 263px;
+  min-height: 50px;
+  padding: 0;
+  border: 0;
+  font-size: 50px;
+  font-weight: 700;
+  line-height: 52px;
+  color: #ffffff;
+  background-color: transparent;
+  transition: opacity 300ms, transform 300ms;
+  cursor: pointer;
+
+  &::before {
+    content: "добавить фильм";
+    position: absolute;
+    bottom: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    z-index: 1;
+    transition: all 0.3s;
+    border-top-width: 1px;
+    border-bottom-width: 1px;
+    border-top-style: solid;
+    border-bottom-style: solid;
+    border-top-color: rgb(255, 255, 255);
+    border-bottom-color: rgb(255, 255, 255);
+    opacity: 0;
+    font-size: 16px;
+    transform: scale(0.1, 1);
+  }
+
+  &:hover::before {
+    opacity: 1;
+    transform: scale(1, 1);
+  }
+
+  &::after {
+    content: "+";
+    position: absolute;
+    bottom: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    z-index: 1;
+    border-radius: 10px;
+    transition: all 0.3s;
+    background-color: rgba(255, 255, 255, 0.1);
+  }
+
+  &:hover::after {
+    opacity: 0;
+    transform: scale(0.1, 1);
+  }
+
+  @include mobile {
+    min-width: 200px;
+    min-height: 40px;
+    margin-bottom: 15px;
+    font-size: 40px;
+    font-weight: 700;
+    line-height: 42px;
+    pointer-events: none;
+  }
+}
