+++ app/components/Header/Header.module.scss
@@ -76,16 +76,16 @@
   line-height: 20px;
 
   @include mobile {
-    font-size: 14px;
     margin-bottom: 10px;
+    font-size: 14px;
   }
 }
 
 .counter {
+  margin-bottom: 10px;
   font-size: 30px;
   font-weight: 700;
   line-height: 32px;
-  margin-bottom: 10px;
 
   @include mobile {
     display: none;
@@ -93,78 +93,10 @@
 }
 
 .counter-mob {
-  font-weight: 500;
   display: none;
+  font-weight: 500;
 
   @include mobile {
     display: inline;
-  }
-}
-
-.add-film-btn {
-  min-width: 263px;
-  min-height: 50px;
-  padding: 0;
-  font-size: 50px;
-  font-weight: 700;
-  line-height: 52px;
-  color: #ffffff;
-  border: 0;
-  background-color: transparent;
-  position: relative;
-  transition: opacity 300ms, transform 300ms;
-  cursor: pointer;
-
-  &::before {
-    content: "добавить фильм";
-    position: absolute;
-    font-size: 16px;
-    bottom: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    z-index: 1;
-    opacity: 0;
-    transition: all 0.3s;
-    border-top-width: 1px;
-    border-bottom-width: 1px;
-    border-top-style: solid;
-    border-bottom-style: solid;
-    border-top-color: rgb(255, 255, 255);
-    border-bottom-color: rgb(255, 255, 255);
-    transform: scale(0.1, 1);
-  }
-
-  &:hover::before {
-    opacity: 1;
-    transform: scale(1, 1);
-  }
-
-  &::after {
-    content: "+";
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    z-index: 1;
-    border-radius: 10px;
-    transition: all 0.3s;
-    background-color: rgba(255, 255, 255, 0.1);
-  }
-
-  &:hover::after {
-    opacity: 0;
-    transform: scale(0.1, 1);
-  }
-
-  @include mobile {
-    min-width: 200px;
-    min-height: 40px;
-    margin-bottom: 15px;
-    font-size: 40px;
-    font-weight: 700;
-    line-height: 42px;
-    pointer-events: none;
   }
 }
